Various changes to improve Java2D rendering
Makes OpenGL displays appear at the same size as Java2D counterparts.
Removes a number of synchronization mechanisms that were superfluous.

+++ msi.gama.core/src/msi/gama/outputs/display/AWTDisplayGraphics.java
@@ -145,7 +145,8 @@ public class AWTDisplayGraphics extends AbstractDisplayGraphics implements Point
 			curHeight = hFromModelUnitsToPixels(attributes.size.getY());
 		}
 		if ( attributes.rotation != null && attributes.rotation.key != null ) {
-			currentRenderer.rotate(Maths.toRad * attributes.rotation.key, curX + curWidth / 2, curY + curHeight / 2);
+			currentRenderer.rotate(Maths.toRad * attributes.rotation.key.doubleValue(), curX + curWidth / 2,
+				curY + curHeight / 2);
 		}
 		currentRenderer.drawImage(img, (int) FastMath.round(curX), (int) FastMath.round(curY), (int) curWidth,
 			(int) curHeight, null);
